Refactor: Updated th eproduction code of Canvas size to handle the decimal values

+++ src/calculateCanvasSize/calculateCanvasSize.ts
@@ -2,5 +2,5 @@ export default function calculateCanvasSize(
   length: string,
   width: string
 ) {
-  return parseInt(length) * parseInt(width);
+  return parseFloat(length) * parseFloat(width);
 }
